Rename src/updater.* -> src/update.*
